Fix for ST plane threshold data type limit

+++ dunetrigger/TriggerSim/TPAlgTools/TPAlgTPCSimpleThreshold.hh
@@ -21,9 +21,9 @@ namespace duneana {
     explicit TPAlgTPCSimpleThreshold(fhicl::ParameterSet const& ps) :
       verbosity_(ps.get<int>("verbosity",0)),
       accum_limit_(ps.get<int>("accum_limit",10)),
-      threshold_tpg_plane0_(ps.get<int>("threshold_tpg_plane0")),
-      threshold_tpg_plane1_(ps.get<int>("threshold_tpg_plane1")),
-      threshold_tpg_plane2_(ps.get<int>("threshold_tpg_plane2"))
+      threshold_tpg_plane0_(ps.get<int16_t>("threshold_tpg_plane0")),
+      threshold_tpg_plane1_(ps.get<int16_t>("threshold_tpg_plane1")),
+      threshold_tpg_plane2_(ps.get<int16_t>("threshold_tpg_plane2"))
     {}
 
     void initialize_channel_state(dunedaq::trgdataformats::channel_t const& channel,
@@ -162,10 +162,9 @@ namespace duneana {
     //configuration parameters
     const int verbosity_;
     const int accum_limit_;
-    const int threshold_tpg_plane0_;
-    const int threshold_tpg_plane1_;
-    const int threshold_tpg_plane2_;
-
+    const int16_t threshold_tpg_plane0_;
+    const int16_t threshold_tpg_plane1_;
+    const int16_t threshold_tpg_plane2_;
 
     //variables for tracking state / hit-finding
     int16_t threshold_;
